requirements.txt
pyTelegramBotAPI==4.8.0
requests==2.28.1
instaloader==4.9.5
python-dotenv==0.21.0
pytest~=7.2.0
pandas~=1.5.2
fpdf~=1.7.2
art~=1.7
SQLAlchemy~=2.0.3
loguru~=0.6.0